+++ eCommerce.Inventory.Infrastructure/Migrations/20260820202633_AddDescriptionToEntities.cs
@@ -1,0 +1,49 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace eCommerce.Inventory.Infrastructure.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddDescriptionToEntities : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "Location",
+                table: "PendingListings");
+
+            migrationBuilder.AddColumn<string>(
+                name: "Description",
+                table: "PendingListings",
+                type: "nvarchar(max)",
+                nullable: true);
+
+            migrationBuilder.AddColumn<string>(
+                name: "Description",
+                table: "InventoryItems",
+                type: "nvarchar(max)",
+                nullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "Description",
+                table: "PendingListings");
+
+            migrationBuilder.DropColumn(
+                name: "Description",
+                table: "InventoryItems");
+
+            migrationBuilder.AddColumn<string>(
+                name: "Location",
+                table: "PendingListings",
+                type: "nvarchar(max)",
+                nullable: false,
+                defaultValue: "");
+        }
+    }
+}